Tests: Added more tests for access-checks
