+++ test/features/errors/notfound.js
@@ -6,7 +6,6 @@
 var assert = require('../../utils/assert.js');
 var preq   = require('preq');
 var server = require('../../utils/server.js');
-var nock = require('nock');
 
 describe('404 handling', function() {
 
@@ -80,92 +79,4 @@ describe('404 handling', function() {
             assert.contentType(e, 'application/problem+json');
         });
     });
-    it('should set page_deleted on deleted page', function() {
-        var apiURI = 'https://en.wikipedia.org/w/api.php';
-        var title = 'TestingTitle';
-        var revision = 12345;
-        var emptyResponse = {'batchcomplete':'','query':{'badrevids':{'12345' :{'revid':'12345'}}}};
-
-        nock.enableNetConnect();
-        var api = nock(apiURI)
-            // The first request should return a page so that we store it.
-        .post('')
-        .reply(200, {
-            'batchcomplete': '',
-            'query': {
-                'pages': {
-                    '11089416': {
-                        'pageid': 11089416,
-                        'ns': 0,
-                        'title': title,
-                        'contentmodel': 'wikitext',
-                        'pagelanguage': 'en',
-                        'touched': '2015-05-22T08:49:39Z',
-                        'lastrevid': 653508365,
-                        'length': 2941,
-                        'revisions': [{
-                            'revid': revision,
-                            'user': 'Chuck Norris',
-                            'userid': 3606755,
-                            'timestamp': '2015-03-25T20:29:50Z',
-                            'size': 2941,
-                            'sha1': 'c47571122e00f28402d2a1b75cff77a22e7bfecd',
-                            'contentmodel': 'wikitext',
-                            'comment': 'Test',
-                            'tags': []
-                        }]
-                    }
-                }
-            }
-        })
-        // Other requests return nothing as if the page is deleted.
-        .post('')
-        .reply(200, emptyResponse)
-        .post('')
-        .reply(200, emptyResponse);
-        // Fetch the page
-        return preq.get({
-            uri: server.config.bucketURL + '/title/' + title,
-            headers: {
-                'cache-control': 'no-cache'
-            }
-        })
-        .then(function(res) {
-            assert.deepEqual(res.status, 200);
-            assert.deepEqual(res.body.items.length, 1);
-            assert.deepEqual(res.body.items[0].rev, revision);
-        })
-        // Now fetch info that it's deleted
-        .then(function() {
-            return preq.get({
-                uri: server.config.bucketURL + '/title/' + title,
-                headers: {
-                    'cache-control': 'no-cache'
-                }});
-        })
-        .then(function() {
-            throw new Error('404 should have been returned for a deleted page');
-        })
-        .catch(function(e) {
-            assert.deepEqual(e.status, 404);
-            assert.contentType(e, 'application/problem+json');
-        })
-        // Getting it by revision id should also return 404
-        .then(function() {
-            return preq.get({uri: server.config.bucketURL + '/revision/' + revision});
-        })
-        .then(function() {
-            throw new Error('404 should have been returned for a deleted page');
-        })
-        .catch(function(e) {
-            assert.deepEqual(e.status, 404);
-            assert.contentType(e, 'application/problem+json');
-        })
-        .then(function() {
-            api.done();
-        })
-        .finally(function() {
-            nock.cleanAll();
-        });
-    })
 });
